(related to bug 452) The allocation manager has two different ways to create reservations:
1 - in the scheduler section, which is the one that we will use in the future
2 - in the backoffice section, where the rules are more general and the users are allowed to do more things. 

Due to this behaviour, the backoffice section ended up with a no working behaviour. Sorry about that. This is fixed with these commits, since it was kind simple to update the code.

Additionally, when putting one event associated to the currently logged in user in the future, is now updated in "My reservations" section of the page.

git-svn-id: file:///tmp/svnroot_only_rec/ReC/trunk@2179 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4