(related to bug 786) Implementation of a single timeout Thread that is notified when the command is sent and it is correctly received. When time for the command has passed, the thread notifies the event listener.
The XML definition was simplified replacing the existing no_data tags for a single no_data like there was already made for the dat_bin timeout. All the time values now are in millis.

The business logic that is going to be executed when timeout occurs is not yet implemented and is only logging the event.

git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@4018 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4