(related to bug 875) Well, this bug fix went terribly:
1 - I downloaded glassfish latest version (3.1), which comes up with Mojarra 2.1.x and installed on my computer
2 - I installed the rec.am application locally and the incorrect behaviour also showed up
3 - I updated to the latest version of primefaces and the initial bug was gone but new ones came up!

Well... after all this time (about 6 hours!), it seems that the version of glassfish that I downloaded and the version of glassfish in production are different and not compatible. Namely, primefaces in the rec.am application is not compatible with the latest Mojarra version (see http://code.google.com/p/primefaces/issues/detail?id=1586). Ouch!

So, the fix was a very small one which, by the way, it was detected yesterday in the first 30 minutes of code analysis. However, that fix, when tested, did not seem to fix the whole thing, thus making me spend more and more time on finding and fixing the bug... :-(

git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@4251 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4