(related to bug 1522) New entity that maps to an Error Message table. Creation of a stateless session bean to add the data to the JMS queue and to process each unit after it is selected from the MDB. It also listens to the timer service. Creation of an MDB to process events from the queue.
To deploy the application, it's necessary to create a queue and a mail session on glassfish with the names mapped in the stateless bean. It's also necessary to create 3 other JNDI Resources, to configure number of supported retries, retry interval and recipients for block.

To avoid conflicts with the Glassfish vendor-specific global jndi names, the  property disable-nonportable-jndi-names should be set to true (since it's EJB 3.1 and a Portable Global JNDI Name is already available, there's no need to use the vendor-specific): set server.ejb-container.property.disable-nonportable-jndi-names="true" in asadmin command line.

git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@6354 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4