(related to bug 121) When building using the target dist.all, with the experiment vpend2m, the attached error occurs (i suppose it happens with every experience).
In the last commit, these lines were added to experimentsbuild.xml:

<replacefilter token="@rec.client.processingmanager.threadpool.coresize@" property="rec.client.processingmanager.threadpool.coresize" />
<replacefilter token="@rec.client.processingmanager.threadpool.maxsize@" property="rec.client.processingmanager.threadpool.maxsize" />
<replacefilter token="@rec.processingmanager.thread.idletime@" property="rec.processingmanager.thread.idletime" />

The replace filter is looking for a property file at "experiments/@{experiment.name}/etc/build.properties", so it won't find these properties since they are only defined in the generic file.

It was fixed to look for these properties in the generic build.properties file. Please check if it's the intended behaviour.

git-svn-id: file:///tmp/svnroot_only_rec/internal/rec-project/rec/trunk@6447 8c4c7310-a39f-45c4-bd0d-b10b2e5d99a4